Fix code review comments and update version.
