+++ edx_rest_framework_extensions/authn/jwt/constants.py
@@ -1,0 +1,5 @@
+"""
+JWT Authentication constants.
+"""
+
+JWT_DELIMITER = '.'
